+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Add setting to disable automatically applying Thermal Nimbus damage.
 - Moved Timber Sentinel to be a hazard instead of an NPC, so it will not provide flanking.
+- Make Protector Tree token respect prototype token settings.
 
 ## 1.3.0 - 2024-11-02 (Pathfinder 2e 6.6.1)
 - Add automation for Timber Sentinel.